move functions one for each file
