+++ lib/functions/put_if_absent_async.dart
@@ -1,0 +1,14 @@
+part of '../ms_map_utils.dart';
+
+  /// Return a value if it exists in map or call [ifAbsent] that's return a new value
+  /// and insert him on map
+Future putIfAbsentAsync<V>(
+    Map map, String key, Future<V> Function() ifAbsent) async {
+  if (map.containsKey(key)) {
+    return map[key];
+  } else {
+    var value = await ifAbsent();
+    map[key] = value;
+    return value;
+  }
+}
